Log malformed error responses from AWS SES
There are times when AWS SES seems to be temporarily unavailable - for
a limited amount of time it returns an error response for perfectly OK
emails. This gets solved on its own when we resend the same message at
a later time. The issue is that error handling fails when trying to read
the error code from the response.

Logging the error to understand what's happening.